Fix dtype: cast FlyDSL fp16 output to match model dtype (bf16)
Made-with: Cursor

+++ vllm/model_executor/layers/fused_moe/gpt_oss_triton_kernels_moe.py
@@ -998,4 +998,4 @@ def _asm_moe_1stage_forward(
          sorted_ids, sorted_expert_ids, sorted_weights.view(-1).contiguous(),
          num_valid_ids, bias_1d, M, model_dim, inter_dim, int(blocks), stream_ptr)
 
-    return out_s2
+    return out_s2.to(hidden_states.dtype)
